backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
kafka-python==2.0.2
scikit-learn==1.5.0
numpy==2.4.4
pandas==2.2.2
joblib==1.4.2
influxdb-client==1.43.0
paho-mqtt==2.0.0
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
